remove unneeded error interceptor
