--- serialization-avro/src/main/java/io/axoniq/dev/samples/serializationavro/api/GiftCardBusinessError.java
@@ -1,11 +1,0 @@
-package io.axoniq.dev.samples.serializationavro.api;
-
-import javax.annotation.Nonnull;
-
-public record GiftCardBusinessError(
-        String name,
-        @Nonnull GiftCardBusinessErrorCode code,
-        String message
-) {
-
-}
