--- serialization-avro/src/main/java/io/axoniq/dev/samples/serializationavro/api/GiftCardBusinessErrorCode.java
@@ -1,12 +1,0 @@
-package io.axoniq.dev.samples.serializationavro.api;
-
-/**
- * Public domain specific error codes
- *
- * @author Stefan Andjelkovic
- */
-public enum GiftCardBusinessErrorCode {
-    NEGATIVE_AMOUNT,
-    INSUFFICIENT_FUNDS,
-    UNKNOWN
-}
