+++ serialization-avro/src/main/java/io/axoniq/dev/samples/serializationavro/api/InsufficientFunds.java
@@ -1,0 +1,12 @@
+package io.axoniq.dev.samples.serializationavro.api;
+
+public class InsufficientFunds extends GiftCardException {
+
+    public InsufficientFunds() {
+        super("amount > remaining value", GiftCardBusinessErrorCode.INSUFFICIENT_FUNDS);
+    }
+
+    public InsufficientFunds(String errorMessage) {
+        super(errorMessage, GiftCardBusinessErrorCode.INSUFFICIENT_FUNDS);
+    }
+}
